style: move SidebarResources
