+++ wondrous-app/components/Common/Sidebar/index.tsx
@@ -1,3 +1,2 @@
 export { default as SideBarComponent } from './Main';
 export { default as EntitySidebar } from './Entity';
-export { default as ResourcesSidebar } from './Resources';
